move and rename utils
