+++ doubleml/double_ml_did.py
@@ -7,8 +7,8 @@ from .double_ml import DoubleML
 from .double_ml_data import DoubleMLData
 from .double_ml_score_mixins import LinearScoreMixin
 
-from ._utils import _dml_cv_predict, _get_cond_smpls, _dml_tune, _trimm
-from ._utils_checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
+from .utils._estimation import _dml_cv_predict, _get_cond_smpls, _dml_tune, _trimm
+from .utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLDID(LinearScoreMixin, DoubleML):
