+++ doubleml/double_ml_did_cs.py
@@ -7,8 +7,8 @@ from .double_ml import DoubleML
 from .double_ml_data import DoubleMLData
 from .double_ml_score_mixins import LinearScoreMixin
 
-from ._utils import _dml_cv_predict, _trimm, _get_cond_smpls_2d, _dml_tune
-from ._utils_checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
+from .utils._estimation import _dml_cv_predict, _trimm, _get_cond_smpls_2d, _dml_tune
+from .utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLDIDCS(LinearScoreMixin, DoubleML):
