+++ doubleml/double_ml_pliv.py
@@ -11,8 +11,8 @@ from .double_ml import DoubleML
 from .double_ml_data import DoubleMLData
 from .double_ml_score_mixins import LinearScoreMixin
 
-from ._utils import _dml_cv_predict, _dml_tune
-from ._utils_checks import _check_finite_predictions
+from .utils._estimation import _dml_cv_predict, _dml_tune
+from .utils._checks import _check_finite_predictions
 
 
 class DoubleMLPLIV(LinearScoreMixin, DoubleML):
