+++ doubleml/double_ml_plr.py
@@ -11,8 +11,8 @@ from .double_ml_data import DoubleMLData
 from .double_ml_score_mixins import LinearScoreMixin
 from .double_ml_blp import DoubleMLBLP
 
-from ._utils import _dml_cv_predict, _dml_tune
-from ._utils_checks import _check_score, _check_finite_predictions, _check_is_propensity
+from .utils._estimation import _dml_cv_predict, _dml_tune
+from .utils._checks import _check_score, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLPLR(LinearScoreMixin, DoubleML):
