+++ doubleml/double_ml_qte.py
@@ -11,9 +11,9 @@ from .double_ml_pq import DoubleMLPQ
 from .double_ml_lpq import DoubleMLLPQ
 from .double_ml_cvar import DoubleMLCVAR
 
-from ._utils import _draw_weights, _default_kde
-from ._utils_resampling import DoubleMLResampling
-from ._utils_checks import _check_score, _check_trimming, _check_zero_one_treatment
+from .utils._estimation import _draw_weights, _default_kde
+from .utils.resampling import DoubleMLResampling
+from .utils._checks import _check_score, _check_trimming, _check_zero_one_treatment
 
 
 class DoubleMLQTE:
