+++ doubleml/tests/test_cv_predict.py
@@ -6,7 +6,7 @@ from sklearn.model_selection import KFold, train_test_split
 from sklearn.linear_model import Lasso, LogisticRegression
 
 from ._utils_dml_cv_predict import _dml_cv_predict_ut_version
-from doubleml._utils import _dml_cv_predict
+from doubleml.utils._estimation import _dml_cv_predict
 
 
 @pytest.fixture(scope='module',
