+++ doubleml/tests/test_lpq.py
@@ -10,7 +10,7 @@ from statsmodels.nonparametric.kde import KDEUnivariate
 
 from ._utils import draw_smpls
 from ._utils_lpq_manual import fit_lpq
-from .._utils import _default_kde
+from ..utils._estimation import _default_kde
 
 
 def custom_kde(u, weights):
