+++ doubleml/utils/__init__.py
@@ -1,7 +1,11 @@
 from .dummy_learners import DMLDummyRegressor
 from .dummy_learners import DMLDummyClassifier
 
+from .resampling import DoubleMLResampling, DoubleMLClusterResampling
+
 __all__ = [
     "DMLDummyRegressor",
     "DMLDummyClassifier",
+    "DoubleMLResampling",
+    "DoubleMLClusterResampling",
 ]
